refactor: file reorganization
